Refactor project structure and logging levels; add ModelManager and ads router
